[fix] 클래스 이름, 메소드 이름 일부 수정
