+++ src/main/java/com/prgrms/rg/web/ridingpost/api/RidingPostInquiryController.java
@@ -1,0 +1,37 @@
+package com.prgrms.rg.web.ridingpost.api;
+
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Slice;
+import org.springframework.data.web.PageableDefault;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.prgrms.rg.domain.ridingpost.application.RidingPostReadService;
+import com.prgrms.rg.domain.ridingpost.model.RidingPostInfo;
+import com.prgrms.rg.domain.ridingpost.model.RidingSearchCondition;
+
+import lombok.RequiredArgsConstructor;
+
+@RestController
+@RequiredArgsConstructor
+public class RidingPostInquiryController {
+	private final RidingPostReadService postReadService;
+
+	@GetMapping("/api/v1/ridingposts/{postId}")
+	public ResponseEntity<RidingPostInfo> getRidingPostDetail(@PathVariable Long postId) {
+		RidingPostInfo ridingPostInfo = postReadService.loadRidingPostInfoById(postId);
+		return ResponseEntity.ok(ridingPostInfo);
+	}
+
+	@GetMapping("api/v1/ridingposts")
+	public ResponseEntity<Slice<RidingPostInfo>> getRidingPostList(
+		@ModelAttribute RidingSearchCondition searchCondition,
+		@PageableDefault Pageable pageable) {
+		var ridingPostSlice = postReadService.loadFilteredRidingPostByCondition(searchCondition, pageable);
+		return ResponseEntity.status(200).body(ridingPostSlice);
+	}
+
+}
